Create bun.yml
This github action will setup bun and its dependencies when a pull or push request is initiated to find out whether the bun application environment executed successfully and what version of bun was used. 


I might use this action as a pre-requsite for a /push/pull/merge into the production branch later on the course.